ingestion/voice/requirements.txt
fastapi
uvicorn[standard]
numpy
pydantic
pydantic-settings
ffmpeg-python
aioredis
websockets
python-multipart
# openai-whisper is now installed from GitHub in Dockerfile
